Refactor-8 moved extra util functions to utils.go
